+++ airbyte-integrations/connectors/source-bing-ads/unit_tests/test_source.py
@@ -2,232 +2,43 @@
 # Copyright (c) 2023 Airbyte, Inc., all rights reserved.
 #
 
-from unittest.mock import MagicMock, patch
-
-import pytest
-import source_bing_ads
-from bingads.service_info import SERVICE_INFO_DICT_V13
-from conftest import find_stream
-from helpers import source
-from source_bing_ads.base_streams import Accounts, Campaigns
-
-from airbyte_cdk.models import SyncMode
-from airbyte_cdk.utils import AirbyteTracedException
+from conftest import get_source
 
 
-@patch.object(source_bing_ads.source, "Client")
-def test_streams_config_based(mocked_client, config):
-    streams = source(config).streams(config)
+def test_streams_config_based(config):
+    streams = get_source(config).streams(config)
     assert len(streams) == 77
 
 
-@patch.object(source_bing_ads.source, "Client")
-def test_source_check_connection_ok(mocked_client, config, logger_mock):
-    with patch.object(Accounts, "read_records", return_value=iter([{"Id": 180519267}, {"Id": 180278106}])):
-        assert source(config).check_connection(logger_mock, config=config) == (True, None)
+def test_source_check_connection_ok(config, logger_mock, mock_auth_token, mock_user_query, mock_account_query):
+    source = get_source(config)
+    assert source.check_connection(logger_mock, config=config) == (True, None)
 
 
-@patch.object(source_bing_ads.source, "Client")
-def test_source_check_connection_failed_user_do_not_have_accounts(mocked_client, config, logger_mock):
-    with patch.object(Accounts, "read_records", return_value=[]):
-        connected, reason = source(config).check_connection(logger_mock, config=config)
-        assert connected is False
-        assert (
-            reason.message == "Config validation error: You don't have accounts assigned to this user. Please verify your developer token."
-        )
-
-
-def test_source_check_connection_failed_invalid_creds(config, logger_mock):
-    with patch.object(Accounts, "read_records", return_value=[]):
-        connected, reason = source(config).check_connection(logger_mock, config=config)
-        assert connected is False
-
-
-@patch.object(source_bing_ads.source, "Client")
-def test_campaigns_request_params(mocked_client, config):
-    campaigns = Campaigns(mocked_client, config)
-
-    request_params = campaigns.request_params(stream_slice={"account_id": "account_id"})
-    assert request_params == {
-        "AccountId": "account_id",
-        "CampaignType": "Audience DynamicSearchAds Search Shopping PerformanceMax",
-        "ReturnAdditionalFields": "AdScheduleUseSearcherTimeZone BidStrategyId CpvCpmBiddingScheme DynamicDescriptionSetting DynamicFeedSetting MaxConversionValueBiddingScheme MultimediaAdsBidAdjustment TargetImpressionShareBiddingScheme TargetSetting VerifiedTrackingSetting",
-    }
-
-
-@patch.object(source_bing_ads.source, "Client")
-def test_campaigns_stream_slices(mocked_client, config):
-    campaigns = Campaigns(mocked_client, config)
-    accounts_read_records = iter([{"Id": 180519267, "ParentCustomerId": 100}, {"Id": 180278106, "ParentCustomerId": 200}])
-    with patch.object(Accounts, "read_records", return_value=accounts_read_records):
-        slices = campaigns.stream_slices()
-        assert list(slices) == [
-            {"account_id": 180519267, "customer_id": 100},
-            {"account_id": 180278106, "customer_id": 200},
-        ]
-
-
-def test_adgroups_stream_slices(mock_auth_token, mock_user_query, mock_account_query, requests_mock, config):
+def test_source_check_connection_ok_but_user_do_not_have_accounts(config, logger_mock, mock_auth_token, mock_user_query, requests_mock):
     requests_mock.post(
-        "https://campaign.api.bingads.microsoft.com/CampaignManagement/v13/Campaigns/QueryByAccountId",
+        "https://clientcenter.api.bingads.microsoft.com/CustomerManagement/v13/Accounts/Search",
         status_code=200,
         json={
-            "Campaigns": [
-                {"Id": 1, "LastModifiedTime": "2022-02-02T22:22:22"},
-                {"Id": 2, "LastModifiedTime": "2022-02-02T22:22:22"},
-                {"Id": 3, "LastModifiedTime": "2022-02-02T22:22:22"},
-            ]
+            "Accounts": []
         },
     )
-    ad_groups = find_stream("ad_groups", config)
-    stream_slices = list(ad_groups.retriever.stream_slicer.stream_slices())
-    assert stream_slices == [
-        {"campaign_id": [1], "parent_slice": [{"account_id": 1, "parent_slice": {"account_name": "", "user_id": 1, "parent_slice": {}}}]},
-        {"campaign_id": [2], "parent_slice": [{"account_id": 1, "parent_slice": {"account_name": "", "user_id": 1, "parent_slice": {}}}]},
-        {"campaign_id": [3], "parent_slice": [{"account_id": 1, "parent_slice": {"account_name": "", "user_id": 1, "parent_slice": {}}}]},
-    ]
+    source = get_source(config)
+    connected, reason = source.check_connection(logger_mock, config=config)
+    assert connected is True
 
 
-def test_ads_request_body_data(mock_auth_token, config):
-    ads = find_stream("ads", config)
-    stream_slice = {
-        "campaign_id": [1],
-        "parent_slice": [{"account_id": 1, "parent_slice": {"account_name": "AccountName", "user_id": 1, "parent_slice": {}}}],
-    }
-
-    request_params = ads.retriever.requester.get_request_body_json(stream_slice=stream_slice)
-    assert request_params == {
-        "AdTypes": ["Text", "Image", "Product", "AppInstall", "ExpandedText", "DynamicSearch", "ResponsiveAd", "ResponsiveSearch"],
-        "ReturnAdditionalFields": "ImpressionTrackingUrls,Videos,LongHeadlines",
-    }
-
-
-def test_ads_stream_slices(mock_auth_token, mock_user_query, mock_account_query, requests_mock, config):
+def test_source_check_connection_failed_invalid_creds(config, logger_mock, mock_auth_token, mock_user_query, requests_mock):
     requests_mock.post(
-        "https://campaign.api.bingads.microsoft.com/CampaignManagement/v13/Campaigns/QueryByAccountId",
-        status_code=200,
-        json={
-            "Campaigns": [
-                {"Id": 1, "LastModifiedTime": "2022-02-02T22:22:22"},
-                {"Id": 2, "LastModifiedTime": "2022-02-02T22:22:22"},
-                {"Id": 3, "LastModifiedTime": "2022-02-02T22:22:22"},
-            ]
-        },
+        "https://clientcenter.api.bingads.microsoft.com/CustomerManagement/v13/Accounts/Search",
+        status_code=401,
+        json={"error": "invalid credentials"},
     )
-    requests_mock.post(
-        "https://campaign.api.bingads.microsoft.com/CampaignManagement/v13/AdGroups/QueryByCampaignId",
-        status_code=200,
-        json={
-            "AdGroups": [
-                {"Id": 1, "LastModifiedTime": "2022-02-02T22:22:22"},
-                {"Id": 2, "LastModifiedTime": "2022-02-02T22:22:22"},
-                {"Id": 3, "LastModifiedTime": "2022-02-02T22:22:22"},
-            ]
-        },
-    )
-    ads = find_stream("ads", config)
-    stream_slices = list(ads.retriever.stream_slicer.stream_slices())
-    assert stream_slices == [
-        {
-            "ad_group_id": [1],
-            "parent_slice": [
-                {
-                    "campaign_id": [1],
-                    "parent_slice": [{"account_id": 1, "parent_slice": {"account_name": "", "user_id": 1, "parent_slice": {}}}],
-                }
-            ],
-        },
-        {
-            "ad_group_id": [2],
-            "parent_slice": [
-                {
-                    "campaign_id": [1],
-                    "parent_slice": [{"account_id": 1, "parent_slice": {"account_name": "", "user_id": 1, "parent_slice": {}}}],
-                }
-            ],
-        },
-        {
-            "ad_group_id": [3],
-            "parent_slice": [
-                {
-                    "campaign_id": [1],
-                    "parent_slice": [{"account_id": 1, "parent_slice": {"account_name": "", "user_id": 1, "parent_slice": {}}}],
-                }
-            ],
-        },
-    ]
+    source = get_source(config)
+    connected, reason = source.check_connection(logger_mock, config=config)
+    assert connected is False
 
 
-@pytest.mark.parametrize(
-    ("stream", "stream_slice"),
-    (
-        (
-            Accounts,
-            {
-                "predicates": {
-                    "Predicate": [
-                        {"Field": "UserId", "Operator": "Equals", "Value": "131313131"},
-                    ]
-                }
-            },
-        ),
-        (Campaigns, {"account_id": "account_id"}),
-    ),
-)
-@patch.object(source_bing_ads.source, "Client")
-def test_streams_full_refresh(mocked_client, config, stream, stream_slice):
-    stream_instance = stream(mocked_client, config)
-    _ = list(stream_instance.read_records(SyncMode.full_refresh, stream_slice))
-    mocked_client.request.assert_called_once()
-
-
-def test_transform(mock_auth_token, config):
-    record = {"AdFormatPreference": "All", "DevicePreference": 0, "EditorialStatus": "ActiveLimited", "FinalAppUrls": None}
-    ads_stream = find_stream("ads", config)
-    expected_record = {
-        "AccountId": 909090,
-        "AdFormatPreference": "All",
-        "AdGroupId": 90909090,
-        "CustomerId": 9090909,
-        "Descriptions": None,
-        "DevicePreference": 0,
-        "EditorialStatus": "ActiveLimited",
-        "FinalAppUrls": None,
-        "FinalMobileUrls": None,
-        "FinalUrls": None,
-        "ForwardCompatibilityMap": None,
-        "Headlines": None,
-        "Images": None,
-        "LongHeadlines": None,
-        "Path1": None,
-        "Path2": None,
-        "TextPart2": None,
-        "TitlePart3": None,
-        "Videos": None,
-    }
-    transformed_record = list(
-        ads_stream.retriever.record_selector.filter_and_transform(
-            all_data=[
-                record,
-            ],
-            stream_state=None,
-            records_schema={},
-            stream_slice={
-                "ad_group_id": [90909090],
-                "extra_fields": {"AccountId": [909090], "CustomerId": [9090909]},
-                "parent_slice": [
-                    {
-                        "campaign_id": [1],
-                        "parent_slice": [{"account_id": 1, "parent_slice": {"account_name": "", "user_id": 1, "parent_slice": {}}}],
-                    }
-                ],
-            },
-        )
-    )[0]
-    assert dict(sorted(transformed_record.items())) == dict(sorted(expected_record.items()))
-
-
-@patch.object(source_bing_ads.source, "Client")
-def test_check_connection_with_accounts_names_config(mocked_client, config_with_account_names, logger_mock):
-    with patch.object(Accounts, "read_records", return_value=iter([{"Id": 180519267}, {"Id": 180278106}])):
-        assert source(config=config_with_account_names).check_connection(logger_mock, config=config_with_account_names) == (True, None)
+def test_check_connection_with_accounts_names_config(config_with_account_names, logger_mock, mock_auth_token, mock_user_query, mock_account_query):
+    source = get_source(config_with_account_names)
+    assert source.check_connection(logger_mock, config=config_with_account_names) == (True, None)
